requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
adamp=0.3.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.5.2=pypi_0
asttokens=3.0.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
blessed=1.20.0=pypi_0
bytecode=0.16.0=pypi_0
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2020.6.20=pypi_0
chardet=3.0.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cmake=3.30.5=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contextlib2=21.6.0=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.10.0=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=4.4.2=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
et-xmlfile=2.0.0=pypi_0
exceptiongroup=1.3.1=pypi_0
executing=2.1.0=pyhd8ed1ab_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
fsspec=2024.9.0=pypi_0
ftfy=6.2.3=pypi_0
future=0.18.2=pypi_0
gensim=4.3.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.36.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
googletrans=3.0.0=pypi_0
gpustat=1.1.1=pypi_0
grad-cam=1.5.4=pypi_0
grpcio=1.67.1=pypi_0
h11=0.16.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2024.12.1=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.25.1=pypi_0
hyperframe=5.2.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.0=pyh41d4057_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jiter=0.9.1=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.2.0=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.11.9=pypi_0
lit=18.1.8=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
medclip=0.0.3=pypi_0
ml-collections=0.1.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=2.5=pypi_0
nltk=3.9.1=pypi_0
nose=1.3.7=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.560.30=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
nystrom-attention=0.0.12=pypi_0
oauthlib=3.2.2=pypi_0
open-clip-torch=2.29.0=pypi_0
openai=2.2.0=pypi_0
opencv-python=4.4.0.42=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
panda=0.3.1=pypi_0
pandas=2.0.0=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=7.2.0=pypi_0
pip=24.2=py38h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_0
plotly=5.24.1=pypi_0
prefetch-generator=1.0.1=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
prompt_toolkit=3.0.48=hd8ed1ab_1
protobuf=5.28.2=pypi_0
psutil=6.0.0=pypi_0
ptflops=0.7.3=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydicom=2.4.4=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=2.4.7=pypi_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.8=2_cp38
pytorchcv=0.0.67=pypi_0
pytz=2024.2=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rfc3986=1.5.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.5=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.3=pypi_0
seaborn=0.12.1=pypi_0
sentence-transformers=3.2.1=pypi_0
sentry-sdk=2.15.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.1.0=py38h06a4308_0
shapely=1.7.1=pypi_0
sinkhorn-knopp=0.2=pypi_0
six=1.15.0=pypi_0
smart-open=7.1.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sns=0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.3=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorly=0.4.5=pypi_0
textaugment=2.0.0=pypi_0
textblob=0.18.0.post0=pypi_0
thop=0.0.31-2005241907=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2020.9.3=pypi_0
timm=0.9.2=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.3=pypi_0
torch=2.0.1=pypi_0
torchaudio=2.0.2=pypi_0
torchmetrics=1.5.2=pypi_0
torchstat=0.0.7=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.1=py38h0a891b7_3
tqdm=4.65.0=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.46.2=pypi_0
triton=2.0.0=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wandb=0.18.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.6=pypi_0
wget=3.2=pypi_0
wheel=0.44.0=py38h06a4308_0
wrapt=1.17.0=pypi_0
xlwt=1.3.0=pypi_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
